adding some integration tests to git ignores
